Refactor remove OutboxEventListener add scheduler for OutboxProcessor triggerProcessing() every minute.
Add swagger annotations for followUser endpoint in SubscriptionController
