+++ src/main/java/school/faang/user_service/event/OutboxEvent.java
@@ -1,8 +1,7 @@
-package school.faang.user_service.entity;
+package school.faang.user_service.event;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
-import jakarta.persistence.EntityListeners;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -12,7 +11,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import school.faang.user_service.listener.OutboxEventListener;
 
 import java.time.LocalDateTime;
 
@@ -23,7 +21,6 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Table(name = "outbox_event")
-@EntityListeners(OutboxEventListener.class)
 public class OutboxEvent {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
