+++ src/test/java/school/faang/user_service/service/GoalServiceTest.java
@@ -14,7 +14,7 @@ import org.mockito.quality.Strictness;
 import school.faang.user_service.outbox.OutboxEventProcessor;
 import school.faang.user_service.dto.GoalDto;
 import school.faang.user_service.dto.GoalFilterDto;
-import school.faang.user_service.entity.OutboxEvent;
+import school.faang.user_service.event.OutboxEvent;
 import school.faang.user_service.entity.Skill;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.goal.Goal;
